refactor(stories,combat): move staked_unsatisfied_beats_for_scene to stories (scene-generic helper)
